feat(APIs): clarification for login and X-Bonita-API-Token usage.
The content type to use was missing and if you use a tool such as Postman, default content type (form-data) will lead to unsuccessful login.

The title "X-Bonita-API-Token cookie" is confusing because the value can be retrieve in a cookie but need to be used in an header.
Rephrasing the title and adding an extra sentence to clarify what need to be done should probably help users to correctly use this feature.