fix(requirements): specific supported OS versions (#826)
Describe the only tested and supported versions. Bonita could eventually work with newer
versions but we cannot commit ourselves on it.
This removal is also done to avoid misunderstanding with end users.

For instance, Bonita 7.6 has been released on December 2017 and Windows Server 2019 on
November 2018: at the Bonita release time, we didn't know that Bonita would work with this
unreleased Windows version.