fix(migration): updates oracle
- Changed : in a first step, run Bonita migration tool to update Bonita platform to version 7.8.4, following the procedure [above](#migrate)
- Changed : run the migration tool to migrate the platform to 7.9+, following the procedure [above](#migrate)
- Removed: the following lines, it's already in the migration procedure : 
install target application server with bonita
etc.
- Changed (I'm not super satistied with that): 
step 4: If you use Oracle, and you are migrating from a version lower than 7.9.0, skip to the section Migrating to Bonita 7.9+ using Oracle below. If you use Oracle and are migrating from version 7.9.0 and higher, add the JDBC driver for your database to bonita-migration/lib. This is the same driver as you have installed in your web server lib directory. You must upgrade to Oracle 12c (12.2.x.y) in order to migrate to 7.9+.
- Fix: the migration full procedure numbering that was broken, based on https://stackoverflow.com/questions/18088955/markdown-continue-numbered-list