fix(7.9): fix broken links
Links to generated pages that don't exist anymore: the pages were previously
generated by the old markdown based solution, and don't exist anymore.
  - api-glossary.adoc
  - connectors-overview.adoc: remove non existing links to connectors list
  - data-handling-overview.adoc
  - first-steps-after-setup.adoc
  - set-up-kpis.adoc
  - software-extensibility.adoc

Fix links still using the markdown in the release-notes

Fix links to page that doesn't exist
  - comparison-of-7-x-and-6-x.adoc: no more subtask page as subtasks don't exist
  anymore in bonita
  - list-of-documents.adoc: link to getting started. WARN in newer doc version,
  a dedicated overview page exist and should then be referenced instead
  - manage-control-in-forms.adoc: remove the link as there is no more 'uid modal
  tutorial'
  - multi-language-pages.adoc.adoc
  change links to tutorial but the content of the target page is not in sync
  with the content of this page (not the same tutorial topic!)
  - variables.adoc: fix link to rest api overview

Fix links to wrong anchors (extra `]]` before the '#' in xref)
tomcat-bundle.adoc
wildfly-bundle.adoc

multi-language-pages.adoc.adoc: change links to tutorial but the content of the
target page is not in sync with the content of this page (not the same tutorial
topic!)

release-notes.adoc: fix link to bcd index of the latest version.
Use antora xref syntax to reference cross component page