fix: use the right Asciidoc syntax for page references
Replace link by xref for internal references
This is the recommended way to do it with Antora, that can check the reference
validity.

Remove mailto links: we don't need them here as they are used as examples and
not to document address to send mail to.

Embed video instead of using links.
Introduce attachments instead of storing files in the `images` directory.
Fix broken link for FakeSMTP download

Additional changes
  - guest-user: fix typo in admonition blocks
  - actors: fix numbered list