doc(context): impact of multi-instance on context
Explain that the UserTask context contains more information in case of a multi-instantiation. The extra data gives the reference of the object in the collection that has been used to create the multi-instance. 
Usually, as a UserTask is created for each element in the collection, we want to display that element in the Form. That is why we need to know the value of the iterator for the specific task instance we are currently displaying (the form is context aware).