fix(2021.2): remove extra migration page
In 2021.2, there was an extra page that has been removed in previous versions. The page moved in 2021.2, and the
propagation content process from 2021.1 seems to fail in managing the change correctly.
This extra page generated FATAL Antora errors:
Duplicate alias: 2021.2@bonita::migrate-from-an-earlier-version-of-bonita-bpm.adoc (specified as: ROOT:migrate-from-an-earlier-version-of-bonita-bpm.adoc)
source: modules/version-update/pages/update-overview.adoc in https://github.com/bonitasoft/bonita-doc.git (ref: 2021.2)