fix(2023.2): correctly generate links to the javadoc v9.0 (#2415)
Correctly defined the AsiiDoc attribute in antora.yml. The value must be
put inside quote, otherwise it is considered as a number and the decimal
part is removed. In that case, the used value is 9 instead of 9.0 and
the javadoc links are broken.

### For reviewers

You can find a link to the Javadoc in
* with the issue:
https://documentation.bonitasoft.com/bonita/2023.2/api/api-index#_java_api
* with the fix:
https://bonitasoft-bonita-doc-build_preview-pr-2415.surge.sh/bonita/latest/api/api-index#_java_api