ci: make "contribution checks" work on forked repo PR (#2664)
The workflow now runs on `pull_request_target` events.
There is no security issue here. The checks are done only on the updated
file of the PR without doing tool installation, cache update or branch
checkout. Only the GitHub API is used.

Using this event allows to create PR comment when the PR is created from
a forked repository.