refactor: update 2 links
The link to the article about the "conversion from Subversion to Git" involved a redirect.
For a more direct access, use the new URL of the article.

The link to a Groovy article sent to a site using a wrong certificate that prevented to access to the article:
   x509: certificate is not valid for any names, but wanted to match mrhaki.blogspot.com
Use a new URL of a site with a valid certificate.