feat: improve how the `la-deployer` version is documented
Compatibility matrix: only reference the la-deployer 1.1.x which can manage all supported versions of Bonita (and some older versions).

In the example demonstrating the usage of the deployer, use a specific version to ensure that the documented command can
be copied and used as it successfully.

The version is stored in an AsciiDoc attribute to ease maintenance: it can be used in various pages and updated easily
in the next versions of the documentation.