fix(2025.1): use valid links to GitHub
The 2025.1 is under development, so we cannot use the `bonitaTechnicalVersion` AsciiDoc attribute in links as there is no release related to this version for now.
So, introduce a new bonitaTagOrBranch AsciiDoc attribute to be able to use the name of the dev branch. Its value could be updated when releasing the beta, RC or GA version.